refactor(SDK): Move the structs to a separate file
